Passing 4/4 test cases.
So far it looks like the file is correctly accounting for leap years on the entries I have tried. Additionally it will now print "No data available" if the date entered is after all of the holidays in the file being searched. Lastly a global constant was added where you can edit the file name that the program is searching for holiday data in without having to find the entry spot further down in the code.